+++ spa-routing-core/src/main/kotlin/com/sparouting/contract/SpaRouteParameterType.kt
@@ -1,7 +1,0 @@
-package com.sparouting.contract
-
-enum class SpaRouteParameterType {
-  STRING,
-  INT,
-  UUID
-}